requirements.txt
beautifulsoup4
celery
Faker
Flask
Flask-Cors
Flask-Login
Flask-SQLAlchemy
lxml
Pillow
PyExecJS
PyMySQL
requests
selenium
WTForms
xmltodict
